// src/types/api.d.ts
export interface components {
    schemas: {
// ...
        User: {
            /** Format: uuid */
            id: string;
            /** Format: email */
            email: string;
            first_name: string;
            last_name: string;
            /** Format: uri */
            avatar_url?: string | null;
            status: string;
            /** Format: date-time */
            email_verified_at?: string | null;
            /** Format: date-time */
            last_login_at?: string | null;
            /** Format: date-time */
            created_at: string;
        };
// ...
        AuthSession: {
            access_token: string;
            refresh_token: string;
            /** @constant */
            token_type: "Bearer";
            expires_in: number;
            user: components["schemas"]["User"];
        };
// ...
    };
// ...
}